dashboard: factor out asset list extraction code
It's useful not only during BugReporting creation, so extract it to a
separate function and move to asset_storage.go.
